Vizia/plugins/vizia-edit/requirements.txt
PyQt5>=5.15.0
python-mpv>=1.0.0
numpy>=1.21.0
Pillow>=9.0.0
